requirements.txt
nicegui~=1.4.8
lxml~=5.0.0
tortoise-orm~=0.20.0
